auth/api/build_image/requirements.txt
flask
flask-bcrypt
flask-views
mysql-connector-python
pyjwt
